+++ onchain/app/gen-signatures/GetOpts.hs
@@ -62,6 +62,7 @@ import Plutus.V2.Ledger.Api (
   BuiltinData,
   FromData (fromBuiltinData),
   LedgerBytes (LedgerBytes),
+  PubKeyHash (PubKeyHash),
   TxId (TxId),
   TxOutRef (TxOutRef),
  )
@@ -69,6 +70,7 @@ import PlutusTx.Builtins qualified as Builtins
 import System.IO (FilePath)
 import System.IO.Error (userError)
 import TrustlessSidechain.HaskellPrelude
+import TrustlessSidechain.Governance (GovernanceAuthority, mkGovernanceAuthority)
 import TrustlessSidechain.MerkleTree (
   MerkleTree,
  )
@@ -369,6 +371,16 @@ parseSidechainPrivKey = eitherReader OffChain.strToSecpPrivKey
 parseSidechainPubKey :: OptParse.ReadM EcdsaSecp256k1PubKey
 parseSidechainPubKey = eitherReader (fmap OffChain.secpPubKeyToSidechainPubKey . OffChain.strToSecpPubKey)
 
+-- | Parse pub key hash.
+parseGovernanceAuthority :: OptParse.ReadM GovernanceAuthority
+parseGovernanceAuthority =
+  eitherReader
+    ( fmap (mkGovernanceAuthority . PubKeyHash . Builtins.toBuiltin)
+        . mapLeft ("Invalid public key hash: " <>)
+        . Base16.decode
+        . Char8.pack
+    )
+
 -- | parses the previous merkle root as a hex encoded string
 parsePreviousMerkleRoot :: OptParse.ReadM LedgerBytes
 parsePreviousMerkleRoot =
@@ -579,6 +591,16 @@ sidechainParamsParser = do
         , metavar "TX_ID#TX_IDX"
         , help "Input UTxO to be spent with the first committee hash setup"
         ]
+
+  governanceAuthority <-
+    option parseGovernanceAuthority $
+      mconcat
+        [ short 'g'
+        , long "governance-authority"
+        , metavar "PUB_KEY_HASH"
+        , help "Public key hash of governance authority"
+        ]
+
   (thresholdNumerator, thresholdDenominator) <-
     option parseThreshold $
       mconcat
@@ -586,6 +608,7 @@ sidechainParamsParser = do
         , metavar "UINT/UINT"
         , help "Threshold ratio for the required number of signatures"
         ]
+
   pure SidechainParams {..}
 
 signingKeyFileParser :: OptParse.Parser FilePath