+++ onchain/app/serialise/Main.hs
@@ -27,6 +27,7 @@ import System.FilePath qualified as FilePath
 import System.IO (FilePath, Handle, print)
 import System.IO qualified as IO
 import System.IO.Error qualified as Error
+import Test.TrustlessSidechain.DummyMintingPolicy as DummyMintingPolicy
 import TrustlessSidechain.CandidatePermissionMintingPolicy qualified as CandidatePermissionMintingPolicy
 import TrustlessSidechain.CheckpointValidator qualified as CheckpointValidator
 import TrustlessSidechain.CommitteeCandidateValidator qualified as CommitteeCandidateValidator
@@ -35,6 +36,7 @@ import TrustlessSidechain.CommitteePlainSchnorrSecp256k1ATMSPolicy qualified as 
 import TrustlessSidechain.DistributedSet qualified as DistributedSet
 import TrustlessSidechain.FUELMintingPolicy qualified as FUELMintingPolicy
 import TrustlessSidechain.HaskellPrelude
+import TrustlessSidechain.FUELProxyPolicy qualified as FUELProxyPolicy
 import TrustlessSidechain.MerkleRootTokenMintingPolicy qualified as MerkleRootTokenMintingPolicy
 import TrustlessSidechain.MerkleRootTokenValidator qualified as MerkleRootTokenValidator
 import TrustlessSidechain.PoCECDSA qualified as PoCECDSA
@@ -43,7 +45,9 @@ import TrustlessSidechain.PoCReferenceInput qualified as PoCReferenceInput
 import TrustlessSidechain.PoCReferenceScript qualified as PoCReferenceScript
 import TrustlessSidechain.PoCSchnorr qualified as PoCSchnorr
 import TrustlessSidechain.PoCSerialiseData qualified as PoCSerialiseData
+import TrustlessSidechain.ScriptCache qualified as ScriptCache
 import TrustlessSidechain.UpdateCommitteeHash qualified as UpdateCommitteeHash
+import TrustlessSidechain.Versioning qualified as Versioning
 
 -- * CLI parsing
 
@@ -197,6 +201,7 @@ main =
   getOpts >>= \options ->
     let plutusScripts =
           [ ("FUELMintingPolicy", FUELMintingPolicy.serialisableMintingPolicy)
+          , ("FUELBurningPolicy", FUELMintingPolicy.serialisableBurningPolicy)
           , ("MerkleRootTokenValidator", MerkleRootTokenValidator.serialisableValidator)
           , ("MerkleRootTokenMintingPolicy", MerkleRootTokenMintingPolicy.serialisableMintingPolicy)
           , ("CommitteeCandidateValidator", CommitteeCandidateValidator.serialisableValidator)
@@ -205,6 +210,11 @@ main =
           , ("CommitteeHashValidator", UpdateCommitteeHash.serialisableCommitteeHashValidator)
           , ("CheckpointValidator", CheckpointValidator.serialisableCheckpointValidator)
           , ("CheckpointPolicy", CheckpointValidator.serialisableCheckpointPolicy)
+          , ("ScriptCache", ScriptCache.serialisableScriptCache)
+            -- Versioning System
+          , ("VersionOraclePolicy", Versioning.serialisableVersionOraclePolicy)
+          , ("VersionOracleValidator", Versioning.serialisableVersionOracleValidator)
+          , ("FUELProxyPolicy", FUELProxyPolicy.serialisableFuelProxyPolicy)
           , -- ATMS schemes
             ("CommitteePlainEcdsaSecp256k1ATMSPolicy", CommitteePlainEcdsaSecp256k1ATMSPolicy.serialisableMintingPolicy)
           , ("CommitteePlainSchnorrSecp256k1ATMSPolicy", CommitteePlainSchnorrSecp256k1ATMSPolicy.serialisableMintingPolicy)
@@ -222,6 +232,8 @@ main =
           , ("PoCSerialiseData", PoCSerialiseData.serialisablePoCSerialiseData)
           , ("PoCECDSA", PoCECDSA.serialisableValidator)
           , ("PoCSchnorr", PoCSchnorr.serialisablePolicy)
+          , -- Validators for FUEL Proxy tests
+            ("DummyMintingPolicy", DummyMintingPolicy.serialisableDummyMintingPolicy)
           ]
         plutusScriptsDotPlutus = fmap (Bifunctor.first (FilePath.<.> "plutus")) plutusScripts
      in case options of