+++ onchain/test/Main.hs
@@ -1,6 +1,7 @@
 module Main (main) where
 
 import Test.Tasty (TestTree, defaultMain, testGroup)
+import Test.TrustlessSidechain.DummyMintingPolicy () -- force compilation
 import Test.TrustlessSidechain.MerkleProofSerialisation qualified as MerkleProofSerialisation
 import Test.TrustlessSidechain.MerkleTree qualified as MerkleTree
 import Test.TrustlessSidechain.MultiSig qualified as MultiSig